+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased] - 2022-05-20
-- Fixed: invalid field name added to dca in DatabaseUtil
+- Fixed: [DatabaseUtil] invalid field name added to dca
+- Fixed: [FormUtil] Warnings in php 8
 
 ## [2.216.0] - 2022-05-17
 - Added: ArrayUtil::insertAfterKey() ([#47])